requirements.txt
# AlphaFactory OS - Python Dependencies (CORRECTED)
# Version: 1.0.1
# Compatible with Python 3.11

# ============================================================================
# CORE DATA & ANALYSIS
# ============================================================================
pandas==2.1.3
numpy>=1.26.2,<3.0.0
# Note: Allow numpy 2.x (TA-Lib needs it)
scipy==1.11.4

# ============================================================================
# BACKTESTING & PORTFOLIO ANALYSIS
# ============================================================================
vectorbt==0.26.2

# ============================================================================
# TECHNICAL ANALYSIS
# ============================================================================
# ta-lib already installed via wheel
# Note: Use >= instead of ==0.3.14b0 (beta not on PyPI)

# ============================================================================
# BROKER API - INTERACTIVE BROKERS
# ============================================================================
ib-insync==0.9.86

# ============================================================================
# DATA SOURCES
# ============================================================================
polygon-api-client==1.13.2
alpha-vantage==2.3.1
yfinance==0.2.32

# ============================================================================
# DATABASE
# ============================================================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# ============================================================================
# MACHINE LEARNING (Month 7+ Shadow Mode)
# ============================================================================
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# ============================================================================
# WEB DASHBOARD & VISUALIZATION
# ============================================================================
dash==2.14.2
plotly==5.18.0

# ============================================================================
# ALERTING & NOTIFICATIONS
# ============================================================================
python-telegram-bot==20.7

# ============================================================================
# CONFIGURATION & UTILITIES
# ============================================================================
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0

# ============================================================================
# API CLIENTS
# ============================================================================
requests==2.31.0

# ============================================================================
# TESTING & QUALITY
# ============================================================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

black==23.11.0
flake8==6.1.0
mypy==1.7.1

# ============================================================================
# PERFORMANCE OPTIMIZATION
# ============================================================================
numba==0.58.1
joblib==1.3.2

# ============================================================================
# DATETIME & TIMEZONE
# ============================================================================
pytz==2023.3.post1
python-dateutil==2.8.2

# ============================================================================
# UTILITIES
# ============================================================================
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0